Rename REST content-type utilities to code-example-utils (#56894)
